+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/TransferDao.java
@@ -1,5 +1,8 @@
 package com.techelevator.tenmo.dao;
 
+import com.techelevator.tenmo.model.Account;
+import com.techelevator.tenmo.model.User;
+
 import java.math.BigDecimal;
 
 public interface TransferDao {
@@ -8,7 +11,8 @@ public interface TransferDao {
     String createTransfer(String sendingUser, String recipientUser, BigDecimal transferAmount);
 
     // Returns the response Pending, Approved or Denied
-    String getTransfer(Long transferId);
+    String getTransfer(int transferId);
+
 
 
 }
